+++ common/CRC16.h
@@ -1,6 +1,7 @@
 #ifndef _CRC16_H
 #define _CRC16_H
+#include "types.h"
 
-unsigned long CRC16(const unsigned char *buf, int size, int key);
+uint16 CRC16(const unsigned char *buf, int size, int key);
 
 #endif
